refactor: share AI summary helper
